+++ vector/src/main/java/im/vector/Matrix.java
@@ -69,7 +69,6 @@ import im.vector.activity.KeysBackupManageActivity;
 import im.vector.activity.SplashActivity;
 import im.vector.analytics.MetricsListenerProxy;
 import im.vector.push.PushManager;
-import im.vector.services.EventStreamService;
 import im.vector.store.LoginStorage;
 import im.vector.tools.VectorUncaughtExceptionHandler;
 import im.vector.ui.badge.BadgeProxy;