+++ vector/src/main/java/im/vector/activity/LoginActivity.java
@@ -98,10 +98,7 @@ import im.vector.push.fcm.FcmHelper;
 import im.vector.receiver.VectorRegistrationReceiver;
 import im.vector.receiver.VectorUniversalLinkReceiver;
 import im.vector.repositories.ServerUrlsRepository;
-import im.vector.services.EventStreamService;
 import im.vector.ui.badge.BadgeProxy;
-import im.vector.ui.themes.ActivityOtherThemes;
-import im.vector.ui.themes.ThemeUtils;
 import im.vector.util.PhoneNumberUtils;
 import im.vector.util.ViewUtilKt;
 
@@ -480,6 +477,7 @@ public class LoginActivity extends MXCActionBarActivity implements RegistrationM
 
         // already registered
         if (hasCredentials()) {
+            /*
             if (null != intent && (intent.getFlags() & Intent.FLAG_ACTIVITY_BROUGHT_TO_FRONT) == 0) {
                 Log.d(LOG_TAG, "## onCreate(): goToSplash because the credentials are already provided.");
                 goToSplash();
@@ -492,6 +490,9 @@ public class LoginActivity extends MXCActionBarActivity implements RegistrationM
                     Log.d(LOG_TAG, "## onCreate(): close the login screen because it is a temporary task");
                 }
             }
+            */
+            Log.d(LOG_TAG, "## onCreate(): goToSplash because the credentials are already provided.");
+            goToSplash();
 
             finish();
             return;